third_party/blink/web_tests/platform/generic/fast/forms/select/option-value-trim-html-spaces-expected.txt

Test for space striping .value attribute of OPTION element

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS o1.value is "text with extra while spaces"
PASS o2.value is "text"
PASS o3.value is " test text "
PASS o4.value is "test  text"
PASS successfullyParsed is true

TEST COMPLETE

text with extra while spaces
text
 test text 
test  text
